+++ src/test/java/fi/testbed2/TestModule.java
@@ -18,7 +18,7 @@ public class TestModule extends MainModule {
         // Same as in MainModule
         bind(DialogBuilder.class).to(DefaultDialogBuilder.class);
         bind(MunicipalityService.class).to(InlineMunicipalityService.class);
-        bind(LocationService.class).to(NetworkLocationService.class);
+        bind(LocationService.class).to(PreferenceBasedLocationService.class);
         bind(CoordinateService.class).to(MercatorCoordinateService.class);
         bind(PreferenceService.class).to(DefaultPreferenceService.class);
         bind(BitmapService.class).to(LruCacheBitmapService.class);